Move padding code into .c and .h files.
